Extract P4ResourceMap class from P4InfoManager
