+++ stratum/hal/lib/p4/CMakeLists.txt
@@ -7,6 +7,7 @@
 add_library(stratum_hal_lib_p4_o OBJECT
     p4_info_manager.cc
     p4_info_manager.h
+    p4_resource_map.h
     utils.cc
     utils.h
 )
